+++ src/components/Type.ts
@@ -1,0 +1,2 @@
+export type rowType = { date: number; day: string; foodCost: number | null; fixedCost: number | null };
+export type dayOfWeekType = { id: number; name: string; color: string };
