--- src/components/Type.vue
@@ -1,4 +1,0 @@
-<script lang="ts">
-  export type rowType = { date: number; day: string; foodCost: number | null; fixedCost: number | null };
-  export type dayOfWeekType = { id: number; name: string; color: string };
-</script>
